requirements.txt
ultralytics>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=8.3.0
PyYAML>=6.0
tqdm>=4.62.0
opencv-python>=4.5.0 